Enable downloading binaries and containers
The *download_run_once* kubespray config option allows to download
container images only once then push to cluster nodes and the
*download_localhost* option makes the installer node a delegate for
pushing images while running the deployment with ansible.

These two options pretends to reduce the amount of traffic during
the deployment time of the Kubernetes cluster.

Issue-ID: MULTICLOUD-425

+++ deployments/build.sh
@@ -35,7 +35,9 @@ function _cleanup {
     echo "Cleaning previous execution"
     docker-compose kill
     image=$(grep "image.*k8plugin" docker-compose.yml)
-    docker images ${image#*:} -q | xargs docker rmi -f
+    if [[ -n ${image} ]]; then
+        docker images ${image#*:} -q | xargs docker rmi -f
+    fi
     docker ps -a --filter "status=exited" -q | xargs docker rm
 }
